Move lifecycle button handling into wifi module
